moving DateTimeOffsetExtensions to OpenTelemetry.Internal, adding tests to projects
