+++ src/OpenTelemetry.Exporter.ZPages/Implementation/ZPagesSpans.cs
@@ -18,6 +18,9 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Timers;
+#if NET452
+using OpenTelemetry.Internal;
+#endif
 
 namespace OpenTelemetry.Exporter.ZPages.Implementation
 {
