+++ src/OpenTelemetry/Internal/DateTimeOffsetExtensions.net452.cs
@@ -16,7 +16,7 @@
 #if NET452
 using System;
 
-namespace OpenTelemetry.Exporter.Prometheus
+namespace OpenTelemetry.Internal
 {
     internal static class DateTimeOffsetExtensions
     {
@@ -25,8 +25,19 @@ namespace OpenTelemetry.Exporter.Prometheus
         private const int DaysPer100Years = (DaysPer4Years * 25) - 1;  // 36524
         private const int DaysPer400Years = (DaysPer100Years * 4) + 1; // 146097
         private const int DaysTo1970 = (DaysPer400Years * 4) + (DaysPer100Years * 3) + (DaysPer4Years * 17) + DaysPerYear; // 719,162
+        private const int DaysTo10000 = (DaysPer400Years * 25) - 366;  // 3652059
+
+        private const long TicksPerMillisecond = 10000;
+        private const long TicksPerSecond = TicksPerMillisecond * 1000;
+        private const long TicksPerMinute = TicksPerSecond * 60;
+        private const long TicksPerHour = TicksPerMinute * 60;
+        private const long TicksPerDay = TicksPerHour * 24;
+
         private const long UnixEpochTicks = TimeSpan.TicksPerDay * DaysTo1970; // 621,355,968,000,000,000
+        private const long UnixEpochSeconds = UnixEpochTicks / TimeSpan.TicksPerSecond; // 62,135,596,800
         private const long UnixEpochMilliseconds = UnixEpochTicks / TimeSpan.TicksPerMillisecond; // 62,135,596,800,000
+        private const long MinTicks = 0;
+        private const long MaxTicks = (DaysTo10000 * TicksPerDay) - 1;
 
         public static long ToUnixTimeMilliseconds(this DateTimeOffset dateTimeOffset)
         {
@@ -34,6 +45,28 @@ namespace OpenTelemetry.Exporter.Prometheus
             // the last digit being off by one for dates that result in negative Unix times
             long milliseconds = dateTimeOffset.Ticks / TimeSpan.TicksPerMillisecond;
             return milliseconds - UnixEpochMilliseconds;
+        }
+
+        public static DateTimeOffset FromUnixTimeMilliseconds(long milliseconds)
+        {
+            const long MinMilliseconds = (MinTicks / TimeSpan.TicksPerMillisecond) - UnixEpochMilliseconds;
+            const long MaxMilliseconds = (MaxTicks / TimeSpan.TicksPerMillisecond) - UnixEpochMilliseconds;
+
+            if (milliseconds < MinMilliseconds || milliseconds > MaxMilliseconds)
+            {
+                throw new ArgumentOutOfRangeException(
+                    nameof(milliseconds),
+                    string.Format("milliseconds must be between {0} and {1}", MinMilliseconds, MaxMilliseconds));
+            }
+
+            long ticks = (milliseconds * TimeSpan.TicksPerMillisecond) + UnixEpochTicks;
+            return new DateTimeOffset(ticks, TimeSpan.Zero);
+        }
+
+        public static long ToUnixTimeSeconds(this DateTimeOffset dateTimeOffset)
+        {
+            long seconds = dateTimeOffset.Ticks / TimeSpan.TicksPerSecond;
+            return seconds - UnixEpochSeconds;
         }
     }
 }
